idp-service/src/main/java/io/harness/idp/plugin/repositories/PluginRequestRepositoryCustomImpl.java
/*
 * Copyright 2023 Harness Inc. All rights reserved.
 * Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
 * that can be found in the licenses directory at the root of this repository, also available at
 * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
 */

package io.harness.idp.plugin.repositories;

import io.harness.annotation.HarnessRepo;
import io.harness.annotations.dev.HarnessTeam;
import io.harness.annotations.dev.OwnedBy;
import io.harness.idp.plugin.beans.PluginRequestEntity;

import com.google.inject.Inject;
import java.util.List;
import lombok.AccessLevel;
import lombok.AllArgsConstructor;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.data.mongodb.core.MongoTemplate;
import org.springframework.data.mongodb.core.query.Criteria;
import org.springframework.data.mongodb.core.query.Query;
import org.springframework.data.support.PageableExecutionUtils;

@HarnessRepo
@AllArgsConstructor(access = AccessLevel.PACKAGE, onConstructor = @__({ @Inject }))
@OwnedBy(HarnessTeam.IDP)
public class PluginRequestRepositoryCustomImpl implements PluginRequestRepositoryCustom {
  private MongoTemplate mongoTemplate;

  @Override
  public Page<PluginRequestEntity> findAll(Criteria criteria, Pageable pageable) {
    Query query = new Query(criteria).with(pageable);
    List<PluginRequestEntity> pluginRequestEntityList = mongoTemplate.find(query, PluginRequestEntity.class);
    return PageableExecutionUtils.getPage(pluginRequestEntityList, pageable,
        () -> mongoTemplate.count(Query.of(query).limit(-1).skip(-1), PluginRequestEntity.class));
  }
}
